Move processor.go to the cdc package to avoid circular dependency
